feat(react-template): restructure test directories
Updated Playwright configuration to organize tests into e2e and unit directories.
Introduced network and coverage fixtures for improved testing capabilities.
