+++ packages/react-template/tests/e2e/utils/playwright.globalSetup.ts
@@ -1,13 +1,14 @@
 import type { FullConfig } from '@playwright/test';
 import { CoverageReport } from 'monocart-coverage-reports';
-import { coverageOptions } from '../../mrc.playwright.config';
+import { coverageOptions } from '../../../mrc.playwright.config';
 
-const globalTeardown = async (config: FullConfig) => {
+const globalSetup = (config: FullConfig) => {
   if (config.webServer && coverageOptions.enabled) {
+    console.info('[playwright]  Coverage enabled with monocart');
     const mcr = new CoverageReport(coverageOptions);
-    await mcr.generate();
+    mcr.cleanCache();
   }
 };
 
 // eslint-disable-next-line import/no-default-export
-export default globalTeardown;
+export default globalSetup;
