+++ packages/react-template/tests/e2e/utils/playwright.globalTeardown.ts
@@ -1,14 +1,13 @@
 import type { FullConfig } from '@playwright/test';
 import { CoverageReport } from 'monocart-coverage-reports';
-import { coverageOptions } from '../../mrc.playwright.config';
+import { coverageOptions } from '../../../mrc.playwright.config';
 
-const globalSetup = (config: FullConfig) => {
+const globalTeardown = async (config: FullConfig) => {
   if (config.webServer && coverageOptions.enabled) {
-    console.info('[playwright]  Coverage enabled with monocart');
     const mcr = new CoverageReport(coverageOptions);
-    mcr.cleanCache();
+    await mcr.generate();
   }
 };
 
 // eslint-disable-next-line import/no-default-export
-export default globalSetup;
+export default globalTeardown;
