+++ packages/react-template/tests/unit/demo/ui/components/DemoFooter/Footer.test.tsx
@@ -1,12 +1,12 @@
+import { DemoFooter } from '@Demo/ui/components/DemoFooter';
 import { render } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router';
-import { DemoHeader } from '../DemoHeader';
 
-describe('Header Component', () => {
-  it('should render the Header component correctly', () => {
+describe('Footer Component', () => {
+  it('should render the Footer component correctly', () => {
     const { container } = render(
       <Router>
-        <DemoHeader />
+        <DemoFooter />
       </Router>,
     );
