+++ packages/react-template/tests/unit/demo/ui/components/DemoHeader/Header.test.tsx
@@ -1,12 +1,12 @@
+import { DemoHeader } from '@Demo/ui/components/DemoHeader';
 import { render } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router';
-import { DemoFooter } from '../DemoFooter';
 
-describe('Footer Component', () => {
-  it('should render the Footer component correctly', () => {
+describe('Header Component', () => {
+  it('should render the Header component correctly', () => {
     const { container } = render(
       <Router>
-        <DemoFooter />
+        <DemoHeader />
       </Router>,
     );
