+++ packages/react-template/vitest.config.mts
@@ -28,7 +28,7 @@ export default defineConfig(({ mode }) => {
         'vitest-sonar-reporter': 'sonar-report.xml',
         junit: 'junit-report.xml',
       },
-      include: ['src/**/*.(spec|test|steps).[jt]s?(x)'],
+      include: ['tests/unit/**/*.(spec|test|steps).[jt]s?(x)'],
       maxWorkers: env.CI ? 2 : undefined,
       coverage: {
         enabled: env.CI === 'true',
